vgaarb: Add module param to allow for choosing the boot VGA device
Add module parameter 'bootdev' to the VGA arbiter to allow the user
to choose which PCI device should be selected over any others as the
boot VGA device.

When using a multi-GPU system with one or more GPUs being used in
conjunction with VFIO for passthrough to a virtual machine, if the
VGA arbiter settles on a passthrough GPU as the boot VGA device,
once the VFIO PCI driver claims that GPU, all display output is lost
and the result is blank screens and no VT access.

Signed-off-by: Cal Peake <cp@absolutedigital.net>